separate PRT tests into separate scripts with plots for debugging
