+++ autotest/prt_track_utils.py
@@ -1,0 +1,50 @@
+import os
+
+import flopy
+import numpy as np
+
+
+def get_track_dtype(path: os.PathLike):
+    """Get the dtype of the track data recarray from the ascii header file."""
+
+    hdr_lns = open(path).readlines()
+    hdr_lns_spl = [[ll.strip() for ll in l.split(",")] for l in hdr_lns]
+    return np.dtype(list(zip(hdr_lns_spl[0], hdr_lns_spl[1])))
+
+
+def check_track_data(
+    track_bin: os.PathLike,
+    track_hdr: os.PathLike,
+    track_csv: os.PathLike,
+):
+    """Check that track data written to binary, CSV, and budget files are equal."""
+
+    # get dtype from ascii header file
+    dt = get_track_dtype(track_hdr)
+
+    # read output files
+    data_bin = np.fromfile(track_bin, dtype=dt)
+    data_csv = np.genfromtxt(track_csv, dtype=dt, delimiter=",", names=True)
+    if len(data_csv.shape) == 0:
+        # https://stackoverflow.com/a/24943993/6514033
+        data_csv = np.array([data_csv])
+
+    assert (
+        data_bin.shape == data_csv.shape
+    ), f"Binary and CSV track data shapes do not match: {data_bin.shape} != {data_csv.shape}"
+
+    # check particle tracks written to all output files are equal
+    # check each column separately to avoid:
+    # TypeError: The DType <class 'numpy._FloatAbstractDType'> could not be promoted by <class 'numpy.dtype[void]'>
+    for k in data_bin.dtype.names:
+        assert np.allclose(data_bin[k], data_csv[k], equal_nan=True)
+
+    # make sure columns all have values in the expected range
+    assert all(data_bin["iprp"] >= 1)
+    assert all(data_bin["irpt"] >= 1)
+    assert all(data_bin["kper"] >= 1)
+    assert all(data_bin["kstp"] >= 1)
+    # assert all(data_bin["ilay"] >= 1)
+    assert all(data_bin["icell"] >= 1)
+    # assert all(data_bin["istatus"] >= 0)
+    assert all(data_bin["ireason"] >= 0)
