+++ autotest/test_prt_Particle.F90
@@ -48,42 +48,47 @@ contains
     call partlist%allocate_arrays(npartmax1, levelMin, levelMax, mempath)
 
     ! check initial array sizes
+    call check(error, size(partlist%iprp) == npartmax1)
+    call check(error, size(partlist%irpt) == npartmax1)
+    call check(error, size(partlist%istopweaksink) == npartmax1)
+    call check(error, size(partlist%istopzone) == npartmax1)
+    call check(error, size(partlist%iTrackingDomain, 1) == npartmax1)
+    call check(error, size(partlist%iTrackingDomain, 2) == nlevels)
+    call check(error, size(partlist%iTrackingDomainBoundary, 1) == npartmax1)
+    call check(error, size(partlist%iTrackingDomainBoundary, 2) == nlevels)
+    ! call check(error, size(partlist%icu) == npartmax1)
+    ! call check(error, size(partlist%ilay) == npartmax1)
+    call check(error, size(partlist%izone) == npartmax1)
+    call check(error, size(partlist%istatus) == npartmax1)
     call check(error, size(partlist%x) == npartmax1)
     call check(error, size(partlist%y) == npartmax1)
     call check(error, size(partlist%z) == npartmax1)
     call check(error, size(partlist%trelease) == npartmax1)
     call check(error, size(partlist%tstop) == npartmax1)
     call check(error, size(partlist%ttrack) == npartmax1)
-    call check(error, size(partlist%istopweaksink) == npartmax1)
-    call check(error, size(partlist%istopzone) == npartmax1)
-    call check(error, size(partlist%istatus) == npartmax1)
-    call check(error, size(partlist%irpt) == npartmax1)
-    call check(error, size(partlist%iprp) == npartmax1)
-    call check(error, size(partlist%iTrackingDomain, 1) == npartmax1)
-    call check(error, size(partlist%iTrackingDomain, 2) == nlevels)
-    call check(error, size(partlist%iTrackingDomainBoundary, 1) == npartmax1)
-    call check(error, size(partlist%iTrackingDomainBoundary, 2) == nlevels)
     if (allocated(error)) return
 
     ! resize particle arrays
     call partlist%reallocate_arrays(npartmax2, mempath)
 
     ! check that arrays have been resized
+    call check(error, size(partlist%iprp) == npartmax2)
+    call check(error, size(partlist%irpt) == npartmax2)
+    call check(error, size(partlist%istopweaksink) == npartmax2)
+    call check(error, size(partlist%istopzone) == npartmax2)
+    call check(error, size(partlist%iTrackingDomain, 1) == npartmax2)
+    call check(error, size(partlist%iTrackingDomain, 2) == nlevels)
+    call check(error, size(partlist%iTrackingDomainBoundary, 1) == npartmax2)
+    call check(error, size(partlist%iTrackingDomainBoundary, 2) == nlevels)
+    ! call check(error, size(partlist%icu) == npartmax2)
+    ! call check(error, size(partlist%ilay) == npartmax2)
+    call check(error, size(partlist%istatus) == npartmax2)
     call check(error, size(partlist%x) == npartmax2)
     call check(error, size(partlist%y) == npartmax2)
     call check(error, size(partlist%z) == npartmax2)
     call check(error, size(partlist%trelease) == npartmax2)
     call check(error, size(partlist%tstop) == npartmax2)
     call check(error, size(partlist%ttrack) == npartmax2)
-    call check(error, size(partlist%istopweaksink) == npartmax2)
-    call check(error, size(partlist%istopzone) == npartmax2)
-    call check(error, size(partlist%istatus) == npartmax2)
-    call check(error, size(partlist%irpt) == npartmax2)
-    call check(error, size(partlist%iprp) == npartmax2)
-    call check(error, size(partlist%iTrackingDomain, 1) == npartmax2)
-    call check(error, size(partlist%iTrackingDomain, 2) == nlevels)
-    call check(error, size(partlist%iTrackingDomainBoundary, 1) == npartmax2)
-    call check(error, size(partlist%iTrackingDomainBoundary, 2) == nlevels)
     if (allocated(error)) return
 
   end subroutine test_dynamic_allocation
